third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/misc/resource-timing-sizes-redirect-worker-expected.txt

This is a testharness.js-based test.
FAIL PerformanceResourceTiming sizes Fetch with redirect test assert_equals: seenCount expected 5 but got 0
Harness: the test ran to completion.
